server/pyproject.toml
[tool.poetry]
name = "noti-temply-admin"
version = "0.1.0"
description = "Noti Temply Admin Server"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "temply_app" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = "^24.1.0"
faker = "^37.1.0"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
types-python-jose = "^3.3.5"
types-passlib = "^1.7.6"
types-aiofiles = "^23.2.0"
pylint = "^3.3.7"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '\.venv/'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
skip_gitignore = true
skip_glob = ["**/__pycache__/**", "**/.venv/**"]

[tool.flake8]
max-line-length = 100
extend-ignore = "E203"
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
    "C0301",  # line-too-long
]

[tool.pylint]
exclude = [".venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
exclude = [".venv"]
